+++ dot/rpc/modules/mocks/core_api.go
@@ -105,14 +105,14 @@ func (_m *MockCoreAPI) InsertKey(kp crypto.Keypair) {
 }
 
 // RegisterRuntimeUpdatedChannel provides a mock function with given fields: ch
-func (_m *MockCoreAPI) RegisterRuntimeUpdatedChannel(ch chan<- runtime.Version) (byte, error) {
+func (_m *MockCoreAPI) RegisterRuntimeUpdatedChannel(ch chan<- runtime.Version) (uint32, error) {
 	ret := _m.Called(ch)
 
-	var r0 byte
-	if rf, ok := ret.Get(0).(func(chan<- runtime.Version) byte); ok {
+	var r0 uint32
+	if rf, ok := ret.Get(0).(func(chan<- runtime.Version) uint32); ok {
 		r0 = rf(ch)
 	} else {
-		r0 = ret.Get(0).(byte)
+		r0 = ret.Get(0).(uint32)
 	}
 
 	var r1 error
@@ -126,11 +126,11 @@ func (_m *MockCoreAPI) RegisterRuntimeUpdatedChannel(ch chan<- runtime.Version) 
 }
 
 // UnregisterRuntimeUpdatedChannel provides a mock function with given fields: id
-func (_m *MockCoreAPI) UnregisterRuntimeUpdatedChannel(id byte) bool {
+func (_m *MockCoreAPI) UnregisterRuntimeUpdatedChannel(id uint32) bool {
 	ret := _m.Called(id)
 
 	var r0 bool
-	if rf, ok := ret.Get(0).(func(byte) bool); ok {
+	if rf, ok := ret.Get(0).(func(uint32) bool); ok {
 		r0 = rf(id)
 	} else {
 		r0 = ret.Get(0).(bool)
